requirements.txt
appnope==0.1.0
astroid==2.3.3
backcall==0.1.0
cachetools==4.0.0
certifi==2019.11.28
chardet==3.0.4
decorator==4.4.2
google-api-core==1.16.0
google-auth==1.11.2
google-cloud-bigquery==1.24.0
google-cloud-core==1.3.0
google-cloud-storage==1.26.0
google-resumable-media==0.5.0
googleapis-common-protos==1.51.0
idna==2.9
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
isort==4.3.21
jedi==0.16.0
jupyter-client==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.18.1
pandas==1.0.1
parso==0.6.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.4
protobuf==3.11.3
ptyprocess==0.6.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pylint==2.4.4
python-dateutil==2.8.1
pytz==2019.3
pyzmq==19.0.0
requests==2.23.0
rsa==4.0
six==1.14.0
tornado==6.0.4
traitlets==4.3.3
urllib3==1.25.8
wcwidth==0.1.8
wrapt==1.11.2
